ref(trace-explorer): Split trace explorer contents into smaller files
This file is getting rather huge, split it out into a few smaller pieces.
